+++ src/main/java/gift/config/InterceptorConfig.java
@@ -14,21 +14,21 @@ public class InterceptorConfig implements WebMvcConfigurer {
         this.authInterceptor = loginInterceptor;
     }
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(authInterceptor)
-            .order(1)
-            .addPathPatterns("/**")
-            .excludePathPatterns(
-                "/login",
-                "/register",
-                "/login/kakao",
-                "/",
-                "/swagger-ui/**",
-                "/swagger-ui.html",
-                "/api-docs/**",
-                "/v3/api-docs/**",
-                "/webjars/**"
-            );
-    }
+//    @Override
+//    public void addInterceptors(InterceptorRegistry registry) {
+//        registry.addInterceptor(authInterceptor)
+//            .order(1)
+//            .addPathPatterns("/**")
+//            .excludePathPatterns(
+//                "/api/members/login",
+//                "/api/members/register",
+//                "/api/members/login/kakao",
+//                "/",
+//                "/swagger-ui/**",
+//                "/swagger-ui.html",
+//                "/api-docs/**",
+//                "/v3/api-docs/**",
+//                "/webjars/**"
+//            );
+//    }
 }
