+++ src/main/java/gift/controller/AuthController.java
@@ -5,8 +5,8 @@ import gift.exception.customException.CustomArgumentNotValidException;
 import gift.exception.customException.CustomDuplicateException;
 import gift.exception.customException.CustomException;
 import gift.exception.customException.PassWordMissMatchException;
-import gift.model.user.UserDTO;
-import gift.model.user.UserForm;
+import gift.model.dto.UserDTO;
+import gift.model.form.UserForm;
 import gift.oauth.KakaoOAuthService;
 import gift.service.JwtProvider;
 import gift.service.UserService;
