+++ src/main/java/gift/controller/OrderController.java
@@ -3,8 +3,8 @@ package gift.controller;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import gift.exception.ErrorCode;
 import gift.exception.customException.AccessDeniedException;
-import gift.model.order.OrderDTO;
-import gift.model.order.OrderForm;
+import gift.model.dto.OrderDTO;
+import gift.model.form.OrderForm;
 import gift.service.OrderService;
 import gift.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
