+++ src/main/java/gift/model/entity/Category.java
@@ -1,4 +1,4 @@
-package gift.model.categories;
+package gift.model.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
