+++ src/main/java/gift/model/form/UserForm.java
@@ -1,4 +1,4 @@
-package gift.model.user;
+package gift.model.form;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
