+++ src/main/java/gift/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package gift.repository;
 
-import gift.model.categories.Category;
+import gift.model.entity.Category;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
