+++ src/main/java/gift/service/ItemService.java
@@ -4,11 +4,11 @@ import gift.exception.ErrorCode;
 import gift.exception.customException.CustomArgumentNotValidException;
 import gift.exception.customException.CustomDuplicateException;
 import gift.exception.customException.CustomNotFoundException;
-import gift.model.categories.Category;
-import gift.model.item.Item;
-import gift.model.item.ItemDTO;
-import gift.model.option.Option;
-import gift.model.option.OptionDTO;
+import gift.model.entity.Category;
+import gift.model.entity.Item;
+import gift.model.dto.ItemDTO;
+import gift.model.entity.Option;
+import gift.model.dto.OptionDTO;
 import gift.repository.CategoryRepository;
 import gift.repository.ItemRepository;
 import java.util.List;
