+++ src/main/java/gift/service/OrderService.java
@@ -3,10 +3,10 @@ package gift.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import gift.exception.ErrorCode;
 import gift.exception.customException.CustomNotFoundException;
-import gift.model.item.Item;
-import gift.model.order.Order;
-import gift.model.order.OrderDTO;
-import gift.model.user.User;
+import gift.model.entity.Item;
+import gift.model.entity.Order;
+import gift.model.dto.OrderDTO;
+import gift.model.entity.User;
 import gift.oauth.KakaoApiService;
 import gift.repository.ItemRepository;
 import gift.repository.OrderRepository;
