+++ src/main/java/gift/service/WishListService.java
@@ -2,10 +2,10 @@ package gift.service;
 
 import gift.exception.ErrorCode;
 import gift.exception.customException.CustomNotFoundException;
-import gift.model.item.Item;
-import gift.model.user.User;
-import gift.model.wishList.WishItem;
-import gift.model.wishList.WishListResponse;
+import gift.model.entity.Item;
+import gift.model.entity.User;
+import gift.model.entity.WishItem;
+import gift.model.response.WishListResponse;
 import gift.repository.ItemRepository;
 import gift.repository.UserRepository;
 import gift.repository.WishListRepository;
