+++ src/test/java/gift/service/CategoryServiceTest.java
@@ -2,7 +2,7 @@ package gift.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.model.categories.CategoryDTO;
+import gift.model.dto.CategoryDTO;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
